Added default github token
Its just too much hassle to add the token to 20+ modules
and all of their branches. In the end, its not sensitive information
because the token only grants readonly access to repos the
'silverstripe-issues' user is connected to, it doesn't have
any "authorization scopes" (see http://developer.github.com/v3/oauth/#create-a-new-authorization).