NEW: Added first cut of travis-run command
The travis-run command is designed to be executed from within your module. It will set things
up locally similarly to the set-up that travis uses, so that you can repeat the execution of
the Travis build.

Assuming you have the travis-support module checked out to ~/travis support, you can run it as

~/travis-support/travis-run

Which is the same as this

~/travis-support/travis-run DB=MYSQL CORE_RELEASE=master

It doesn't currently do anything with Behat, and makes a *lot* of assumptions about your
.travis.yml file.  Still, it's a start